requirements.txt
python-vlc
RPi.GPIO
click
